:- updated Design.md with new ideas 	- also added server-data-vanilla/ 		cloned: https://github.com/citizenfx/cfx-server-data.git server-data 		added example default c# projects: 			$ dotnet new cfx-resource
:- new business model attempt…. 
ref: discord -> TDR (@filletkastro)
so i have a demo ready for testing. the death match thing i mean.

but i was testing with someone last night and i realize that i need something 'more' in order to gain attention.

so i am starting a youtube channel for marketing. and i am creating a bunch of different mod types (or game types) to demo on the channel. and then i'll have a link to our discord for people to join if they want to play the mods themselves.

here is a list of initial games, just to gain attention and get people to join the server (i realize that its very hard to get people to just join a straight death match betting type server... ie. we need to rock the boat a little bit first and gain attention) ....

if you are experienced in moderations and admin and running a discord server. i would very much like to partner with you.

like i said, its all about gaining attention. 
and once we get people into our discord (however we do that)... then we can push the real products that we want (ie. your RP server and my death match server)

thoughts?